oncosense/requirements.txt
qiskit>=1.0
qiskit-aer>=0.13
qiskit-machine-learning>=0.7
scikit-learn>=1.3
streamlit>=1.30
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
plotly>=5.18
fpdf2>=2.7
torch>=2.0
torchvision>=0.15
Pillow>=10.0
